+++ views/register.ejs
@@ -1,5 +1,5 @@
 <h3>Register</h3>
-<form action="/register" method="POST"  >
+<form action="/register" method="POST">
     <div>
         <div>
             <label for="username" style="font-weight: bold;">Username</label>
